Finished `primary_left` scene serialization
