+++ integration_tests/examples/demo/outputs/rtp.rs
@@ -16,7 +16,7 @@ use strum::{Display, EnumIter, IntoEnumIterator};
 use tracing::error;
 
 use crate::{
-    inputs::InputHandler,
+    inputs::{filter_video_inputs, InputHandler},
     outputs::{scene::Scene, AudioEncoder, OutputHandler, VideoEncoder, VideoResolution},
     players::OutputPlayer,
     IP,
@@ -435,29 +435,15 @@ impl RtpOutputVideoOptions {
         inputs: &[&dyn InputHandler],
         output_name: &str,
     ) -> serde_json::Value {
-        let input_json = inputs
-            .iter()
-            .filter_map(|input| {
-                if input.has_video() {
-                    let input_name = input.name();
-                    let id = format!("{input_name}_{output_name}");
-                    Some(json!({
-                        "type": "input_stream",
-                        "id": id,
-                        "input_id": input_name,
-                    }))
-                } else {
-                    None
-                }
-            })
-            .collect::<Vec<_>>();
+        let inputs = filter_video_inputs(inputs);
+
         json!({
             "resolution": self.resolution.serialize(),
             "encoder": {
                 "type": self.encoder.to_string(),
             },
             "initial": {
-                "root": self.scene.serialize(&self.root_id, input_json, self.resolution),
+                "root": self.scene.serialize(&self.root_id, &inputs, output_name, self.resolution),
             }
         })
     }
@@ -467,24 +453,10 @@ impl RtpOutputVideoOptions {
         inputs: &[&dyn InputHandler],
         output_name: &str,
     ) -> serde_json::Value {
-        let input_json = inputs
-            .iter()
-            .filter_map(|input| {
-                if input.has_video() {
-                    let input_name = input.name();
-                    let id = format!("{input_name}_{output_name}");
-                    Some(json!({
-                        "type": "input_stream",
-                        "id": id,
-                        "input_id": input_name,
-                    }))
-                } else {
-                    None
-                }
-            })
-            .collect::<Vec<_>>();
+        let inputs = filter_video_inputs(inputs);
+
         json!({
-            "root": self.scene.serialize(&self.root_id, input_json, self.resolution),
+            "root": self.scene.serialize(&self.root_id, &inputs, output_name, self.resolution),
         })
     }
 }
